Supporting dall-e 3 image models
Required an update to the JS SDK, which also brought int a bunch of changes to how chat completions work.

Rewrote the chat completion type usage and data flow to be more in line with how the SDK is expected to be used.

Can generate DALL-E 3 images, but layout is a bit janky